backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
email-validator
cryptography
python-dotenv
python-jose[cryptography]
python-multipart
passlib[bcrypt]
